ml/requirements.txt
requests
python-dotenv
uvicorn
gunicorn
opentelemetry-instrumentation-fastapi
fastapi
pydantic
typing
datetime
pymongo
numpy
pandas
pathlib
contextvars
fastparquet
datamodel-code-generator